requirements.txt
alembic==1.9.3
black==23.1.0
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.3
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.3
